edge-node/pyproject.toml
[project]
name = "edge-node"
version = "0.1.0"
description = "Camera Streaming Node"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "eclipse-zenoh>=1.8.0",
    "opencv-python-headless>=4.13.0.92",
    "python-dotenv>=1.2.2",
]